Fix green filter

+++ lib/filter.py
@@ -43,8 +43,8 @@ def filter_image(original_img, filter_num):
     if filter_num == 5:
         hsv = cv2.cvtColor(rgb_img, cv2.COLOR_BGR2HSV)
 
-        lower_black = np.array([50, 50, 50], dtype="uint16")
-        upper_black = np.array([100, 255, 100], dtype="uint16")
+        lower_black = np.array([50, 50, 50], dtype="uint8")
+        upper_black = np.array([100, 255, 100], dtype="uint8")
         mask = cv2.inRange(hsv, lower_black, upper_black)
 
         img = np.zeros((mask.shape[0], mask.shape[1], 3))
